ARROW-293: [C++] Implement Arrow IO interfaces for operating system files
I started with the code I put together previously for Feather and conformed it to the `arrow::io` API. There's a bunch of Windows compatibility stuff; I left this until we add CI for Windows and can sort this out.

We should also refactor the memory mapped file interfaces to be based on this common code (see ARROW-294).

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #146 from wesm/ARROW-293 and squashes the following commits:

a2653b7 [Wes McKinney] cpplint
d56ef06 [Wes McKinney] Test the rest of ReadableFile methods
43126ca [Wes McKinney] Drafting OS file IO implementations based on Feather implementation. Work on test suite