ARROW-96: Add C++ API documentation
This adds a basic `Doxyfile` that can be used to generate the HTML API documentation as well as a small, initial "Getting Started".

The documentation is not yet deployed anywhere. We can either also use `readthedocs.org` for this (via the `breathe` package) or wait for the restructuring of the website as discussed on the ML and then add this to the "update website" scripts. I'd personally prefer the latter.

Author: Uwe L. Korn <uwelk@xhochy.com>

Closes #271 from xhochy/ARROW-96 and squashes the following commits:

45c98cb [Uwe L. Korn] Add license headers
e7c9849 [Uwe L. Korn] ARROW-96: Add C++ API documentation