ARROW-668: [Python] Box timestamp values as pandas.Timestamp if available, attach tzinfo
I'm not sure how to easily test the behavior if pandas is not present. I created an environment without pandas and added some fixes so that I verify the behavior, but at some point we should create a "no pandas" test suite to see what using pyarrow is like without pandas installed.

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #487 from wesm/ARROW-668 and squashes the following commits:

554a647 [Wes McKinney] Remove cython from requirements.txt
649d28a [Wes McKinney] Box timestamp values as pandas.Timestamp if available, return timezone also if available