ARROW-1366: [Plasma] Define entry point for the plasma store
The method for starting the Plasma store is already documented in https://arrow.apache.org/docs/python/plasma.html. So far it only worked if the store was installed with "make install" from the C++ sources. This makes it also possible to start it if the pyarrow wheels are installed.

Author: Philipp Moritz <pcmoritz@gmail.com>

Closes #975 from pcmoritz/plasma-store-ep and squashes the following commits:

eddc487a [Philipp Moritz] make plasma store entry point private
4c051409 [Philipp Moritz] define entry point for the plasma store