ARROW-905 [Docs] Dockerize document generation
This PR has two parts. First part is to use Docker to generate the API documentation for C, C++, Python and Java. This part is now complete.
The second part is to be able to run the site locally, which is still in progress.

Author: Heimir Sverrisson <heimir.sverrisson@gmail.com>

Closes #1162 from heimir-sverrisson/hs/dockerize_api_docs and squashes the following commits:

f758fdc3 [Heimir Sverrisson] ARROW-905 [Docs] Add Dockerfile for reproducible documentation generation