ARROW-2350: Consolidated RUN step in spark_integration Dockerfile
I am a big fan of this project! In this PR, I propose a small change to the `spark_integration` Dockerfile. See [ARROW-2350](https://issues.apache.org/jira/browse/ARROW-2350) for a full description of what I'm proposing here.

This change cuts the size of the `spark_integration` container from **2.65 GB** to **2.26 GB**.

Author: James Lamb <james.lamb@uptake.com>

Closes #1787 from jameslamb/smaller_docker and squashes the following commits:

7343d8d <James Lamb> ARROW-2350: Consolidated RUN step in spark_integration Dockerfile