ARROW-3473: [Format] Clarify that 64-bit lengths and null counts are permitted, but not recommended
The Arrow metadata was changed from the initial format specification to permit 64-bit array lengths, support for which is provided by the C++ library. This clarifies that 64-bit lengths are permissible, but for best compatibility (e.g. with Java), it is recommended to use 32-bits or less in practice.

see also #2733

Author: Wes McKinney <wesm+git@apache.org>

Closes #2734 from wesm/ARROW-3473 and squashes the following commits:

11b4b08a9 <Wes McKinney> Add note about representing large data sets using chunks
067ec9932 <Wes McKinney> Clarify that 64-bit lengths and null counts are permitted, but not recommended for best inter-language compatibility