ARROW-5187: [Rust] Add ability to convert StructArray to RecordBatch
The CPP version (http://arrow.apache.org/docs/python/generated/pyarrow.StructArray.html?highlight=flatten#pyarrow.StructArray.flatten) returns `Vec<ArrayRef>`, so not sure if we should also do the same.
My justification for returning a `RecordBatch` is that we have the added convenience of not losing the schema of the columns.
I'm using this for reading rows from a SQL database into Arrow data.

Author: Neville Dipale <nevilledips@gmail.com>

Closes #4178 from nevi-me/ARROW-5187 and squashes the following commits:

7afbcfecb <Neville Dipale> make struct_array -> record_batch a conversion trait impl
418bfe1e2 <Neville Dipale> ARROW-5187:  Add StructArray::flatten