ARROW-5635: [C++] Added a Compact() method to Table.
A column in a table may consist of multiple chunks. This PR is proposing a Table.Compact() method that returns a table whose columns are of one chunk, which is the concatenation of the corresponding column's chunks.

This method could be useful if the table is fragmented (after Table.slice(), then Table.concatenate()) while one wants to conduct vectorized computation through to_numpy().

Author: Zhuo Peng <1835738+brills@users.noreply.github.com>

Closes #4598 from brills/compact and squashes the following commits:

6bbd0174a <Zhuo Peng> Rename as CombineChunks and fixed doc.
db640d6cd <Zhuo Peng> Addressed PR comments.
18aa9ebd8 <Zhuo Peng> Added a Compact() method to Table.