ARROW-5257: [Website] Update site to use "official" Apache Arrow logo, add clearly marked links to logo
Header uses an inverted (white print with transparent background) version of the logo because the header has a dark background.

<img width="400" alt="powered-by-header" src="https://user-images.githubusercontent.com/2975928/60135888-9ccadc00-9757-11e9-8469-18f9d09e0387.png">

The proper black logo (with transparent background) is shown on the "powered by" page, which recommends folks to use that logo when referring to Arrow.

<img width="400" alt="powered-by-screenshot" src="https://user-images.githubusercontent.com/2975928/60135811-65f4c600-9757-11e9-8327-692ddbabbace.png">

The header logo is reduced in size to 300px in height, and then the img tag further reduces display to 60px. That combination got the sharpest rendering of the image with a reasonable file size. As discussed on the jira ticket, a better solution would be to use svg, but this works for now.

I also reorganized the powered-by page, putting the projects first and the trademark language last. Given the intended audience of this page, that seems like a more natural order.

Author: Neal Richardson <neal.p.richardson@gmail.com>

Closes #4696 from nealrichardson/correct-logo and squashes the following commits:

481c19149 <Neal Richardson> Add François to committers
cca872150 <Neal Richardson> Move trademark back up
53cd0c406 <Neal Richardson> WIP on getting the right logo