ARROW-5658: [JAVA] Sync schema for VectorSchemaRoot
Resolve JIRA ARROW-5658.

The fundamental problem is that, as data are inserted to the vector (e.g. ListVector), the schema of VectorShemaRoot can be different from the vector structure.

In the server side, the deserialization is based on the schema, which is out of date, so it fails silently.

In this PR, we fix the problem of obsolete schema and make the server print error info explicitly.

Author: liyafan82 <fan_li_ya@foxmail.com>

Closes #4689 from liyafan82/fly_5658 and squashes the following commits:

e7d5865a5 <liyafan82>  Undo throwing exception
3ea77bc76 <liyafan82>  Replace automatic updating schema with throwing an exception
cb9da2032 <liyafan82> Merge branch 'master' into fly_5658
47a776fbe <liyafan82>  Automatically update schema
6d2763848 <liyafan82>  Resolve comments
061e8bc2f <liyafan82>  Fix error log
e8ea49f00 <liyafan82>  Sync schema for VectorSchemaRoot